external modules: sql db functions advances
